backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
google-genai==1.14.0
pinecone==6.0.0
httpx==0.28.1
pydantic==2.11.1
pydantic-settings==2.8.1
